+++ integration/inspect_test.go
@@ -86,7 +86,7 @@ func TestInspect(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
+		t.Run(tt.name, func(_ *testing.T) {
 			resp, err := Inspect(tt.args.volume, tt.args.file, tt.args.encrypt)
 			if tt.expectedError {
 				assert.Nil(err)