+++ pkg/utils/utils_test.go
@@ -81,7 +81,7 @@ func TestDecodeInput(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
+		t.Run(tt.name, func(_ *testing.T) {
 			got, err := DecodeBase64(tt.args.s)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("DecodeBase64() error = %v, wantErr %v", err, tt.wantErr)
@@ -119,7 +119,7 @@ func TestClientIPFromContext(t *testing.T) {
 		},
 	}
 	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
+		t.Run(tt.name, func(_ *testing.T) {
 			if got := ClientIPFromContext(tt.args.ctx); got != tt.want {
 				t.Errorf("ClientIPFromContext() = %v, want %v", got, tt.want)
 			}